MSA-618: Syncs a Simulated garage door device with 2 actual devices, either a tilt or contact sensor and a switch or relay. The simulated device will then control the actual garage door. In addition, the virtual device will sync when the garage door is opened manually, \n It also attempts to double check the door was actually closed in case the beam was crossed.
create a virtual device (see elsewhere for instructions) for device type use the simulated garage door.
You can also create a modified version for yourself I recommend decreasing the runin time to 1 sec, otherwise it takes 6 secs before the door starts closing or opening when you toggle the simulated device icon. (That is what I did)

install this smart app.

select your relay /switch, and actual garage door sensor in the app, Also select the virtual one twice in the app. Once for the sensor (to get the current state), and once for the door control to actually be able to toggle it.

See picts below.

It also optionally can send text or notifications

It also double checks (in 15 seconds) after opening or closing and will resync for some reason if the door didn't really close ie beam crossed, or didn't really open for some reason,