MSA-726: I wrote this app to add some garage-door-specific smarts to my outside lighting.  I wasn't able to find anything that was aware of multiple garage door sensors, and that would keep the lights on for a certain amount of time after the last door was closed.  The use case is as follows:
A garage door goes up while it's dark.  All lights turn on and stay on as long as the door is open, which is useful if you're working outside at night.  If you were actually pulling a car out to leave, you'd want the lights on as you pull out of the driveway, so there is an option to keep them on for a certain time after the door closes.