Final cleanups prior to submission of Pull request
Deleted several commented out sections of code. Cleans up other sections.

There are a few placeholders (commented out) that are awaiting fixes for a problem with settings and dynamic enum lists. This causes some child devices not to be automatically deleted if the SmartApp is used to update the chosen thermostats.