MSA-908: This smart app allows you to integrate a home security system that is using an EnvisaLink controller into the SmartThings platform.  With this smart app the user can use the mobile app to view status as well as arm/disarm their security system.  Additionally the user can setup routines within SmartThings to arm the system at certain times of day, arm the system when they leave the house, etc.
The smart app and child devices have been tested over the last month and has no issues that I am aware of.  I would like to publish this app and its child devices to so others can more easily make use of it.  The full source code for the j64AlarmSystem as well as this smart app can be found on my github at https://github.com/joejarvis64/j64.AlarmServer.