Local foscam with additional features
Tried to get the feature set of the local foscam controller closer to parity with the standard one. Areas still needing some help:

Only tested on 8910W - stripped out most of the logic to try and detect device version when I was trying to narrow down some issues.
No polling - so no way of knowing if alarm is on or off - also cannot currently toggle alarm status.
Removed the setting of presets.
Still not perfect, but it's good enough for me that I'm switching to it and closing my firewall.