Fixes for level
The math ST uses to convert the ST level range (0-100) to the Hue bri
range (1-254) is flawed and causes a number of minor inconsistencies.
For one, if Hue reported a brightness of 1, ST would set the level to 0.
Additionally, setting the level to 1 in ST would set the actual bulb
brightness to 3.