MSA-1332: when a specified contact sensor is closed - verify all other specified open/close contact sensors are also closed - if so, set thermostat back on (to revert changes made by thermostat mode director smartapp (thermostat and doors).
i installed the thermostat mode director smartapp, which works for turning off the thermostat if any contact sensor is open for a specified amount of time - but the smartapp does not include logic to revert the change once the contacts are closed again.

i wrote this smartapp to achieve returning the thermostat to a specified mode when contacts are closed if the thermostat currently isn't set to the desired mode. this reduces the number of push notifications by eliminating unnecessarily changing the thermostat to a mode that it is already set at.