MSA-1395: Smart Family Presence is a smart app with no handlers or device type dependencies.
The presence sensors have little/no configuration, especially when it comes to alerts.  I found lots of room for improvement regarding how those alerts are carried out.  Here are the problems with the original implementation of presence alerts:

1)  Each person gets and individual alert.  If you have a family of 4, each person will receive a *separate* alert for every person.
2)  There are no exceptions.  Even if the family is all together, it still individually alerts that everyone arrived/departed.
3)  Since you can't control which person gets the push alert, #1 and #2 exacerbate the issue by people receiving alerts for themselves.

The Smart Family Presence app lets the user select how and when they receive alerts.  The code is commented to explain the edge cases, but essentially, if the family arrives or departs together, no alert should be sent.  The app tracks state for whether the family was previously apart, so that it knows when and when not to send alerts.

I have found this code extremely fun to write and configure, and I believe others out in the community could benefit from it as I saw very few smart apps related to presence sensors themselves.