Configure ZoneStatus reporting for Device Health and Hub Offline usecases
Previously we were configuring the Temperature attribute for reporting at
5 minute intervals so that we could relatively quickly determine if the
device is offline for Device Health purposes. This changes Temperature
reporting back to the default and instead configures Zone Status reporting
for every 5 minutes. The benefit of this is if the hub is offline when
the device sends a Zone Status Change Notification indicating moisture,
contact or motion, the attribute reporting will get the platform back in
sync with the current state within 5 minutes of the hub coming back online.

https://smartthings.atlassian.net/browse/SSVD-2858