MSA-1491: This app is an improvement to the already great Gentle Wake Up smart app. It adds the ability to define start and end colors and temperatures.
More information here:
https://github.com/gladclef/SmartThingsPublic/wiki/Gentle-Dimmer

Why I decided to make this its own app instead of publishing it as an update to Gentle Wake Up:
https://community.smartthings.com/t/new-features-in-gentle-wake-up-how-to-publish/57220/8

One question I have: do I need to submit the Gentle Wake Up Controller code for submission? It is exactly the same as the existing Gentle Wake Up Controller. I just had to publish it locally to work on the smart app.