Update CentraLite DTHs to reset to factory default
This change adds a call to reset the devices to factory default before
the rest of the normal configuration.  This is to work around a bug in
the CentraLite firmware where the binding table is sometimes not getting
properly cleared through other methods.

This resolves: https://smartthings.atlassian.net/browse/DPROT-186