Read on/off status for bulbs that don't report
Many ZLL bulbs don't report changes to attributes we are interested in.
As a result, we are now manually querying for the status on commands
that could change the status.  In many places this was already done,
however, in a few it was not.  With some upcoming zigbee-library changes
these will be necessary to make sure the state is kept up to date.

This relates to: https://smartthings.atlassian.net/browse/DPROT-227
This relates to: https://smartthings.atlassian.net/browse/DPROT-228