MSA-1733: This application allows a user to set up from 1 to 5 lighting scenes, controlled by momentary switches. The switches could be physical momentary wall switches, or virtual switches created in the IDE.
The lighting scene options include setting color hue and saturation for color-changing bulbs, & lighting level and color temperature for color temperature bulbs (which includes color-changing bulbs as well). A 'scene' can change any or all of the parameters. As an example, 'Movie Scene' could set the lighting level on all lights to 25%, the color temperature for color temp bulbs to 2700k, and leave the colors alone; while 'Full Bright' could set the color temperature to 5500k and brightness to 100% for all bulbs. 'Green' mode could set the hue/saturation of color bulbs to full green, while leaving brightness levels as is, or could set the brightness to a different level at the same time.