MSA-1774: Slickspaces is a web enabled smartapp that allows more advanced integration of smart locks and smart home technology via our web dashboard at slickspaces.com.  It also allows a property manager to easily manage and control multiple locks/hubs/properties from a single interface.
Our goal is to provide the best automation and tools to short term rental property managers to streamline guest check-in, audit access logs and  reduce energy consumption.

All Smartthings oAuth tokens are encrypted and stored in our database using OpenSSL encryption and our database is only accessible on the internal network of the webserver farm, not directly on the internet.  

This is our preliminary version which allows the setting and deleting of lock codes, as well as scheduling them.  We've also enabled hooks to control other popular devices as well as query status.  In the near future we will be moving away from using the smartthings scheduling service (primarily due to your soft limit of 4 schedules per hub) and will just be sending commands for immediate action from our platform.

Our platform goal is simple.  We want to bring smart home technology to the short term rental market.  Part of our business will be to provide pre-configured locks and hubs to customers who subscribe to our service for a monthly fee.  

In order to allow the customer to still have control over their hub and Smartthings account, we very much need to have a published app, otherwise the oAuth handshake becomes a difficult/challenging task to manage for a "self published" app that lives on different users accounts. 

While we see our product as providing an essential service to our customers, it will also often be the first interaction people have with smart home technology.  We anticipate that they will want to connect their own devices and will be a great "gateway" to purchasing smart equipment for their principal residences as well.

Thanks for your consideration

Mathew Hunter
Co-Founder of Slickspaces