MSA-1875: Originally developed by Don Caruanna with minor changes from Sensative. This is intended to be the official/default device handler for Strips by Sensative.
The preference defaults are set to the device defaults. 
(If you are looking for accurate values in the IDE, you have to hit "Done" in preferences once, even if you don't change any settings.)

Preferences with an asterisk have to wait for the next wakeup, automatic or manual. The pending icon will display if there are changes pending and say "Synced" otherwise. If you revert your preferences to values that match the device before the next wakeup, it will go back to "Synced". Only changed preferences are pushed to the device.

The actual wakeup by the device is approximately 90% of the wakeupinterval setting. So the observed wakeup interval will be a little less than the chosen one. Also, the wakeupinterval has a minimum of 1800 and a maximum of 86400.

A tamper notification is sent in 2 cases:
Putting a magnet on the round edge when there is a magnet on the square one triggers tamper
During a Manual Wake Up, a tamper is sent before the Node Info.
The tamper event time can be cleared by just tapping on the time itself. Tamper events are reported so they can be seen historically in the GUI.