MSA-2030: A simple app that monitors the level of water and enforces an "off" state on a switch when no water is present while sending a notification.
For use with a Christmas tree, place a water sensor in the tree stand and fill with water. Then, plug the tree's lights into a smart switch. Configure the app's settings and rest assured your tree's lights won't light if the water is empty. This may help prevent lighting your tree if it is drying out.

Warning: This app does not claim to prevent danger but only notifies the user of a potentially hazardous situation.