Have zigbee setColor honor level settings
The ColorControl.setColor command takes a color map as an argument which
can include a level.  This level was not being honored by most of the
zigbee color bulbs.  This shows up as an issue when using the lighting
automation to set color and level, the level portion is not working.

This resolves: https://smartthings.atlassian.net/browse/DVCSMP-3133