Fix problem setting color temperature on Sengled Element Plus
The Sengled Element Plus seems to ignore the Move to Color Temperature
command when the transition time is 0x0000 but will execute the command
correctly when the value is 0x0001 or greater. The setColorTemperature
command now checks the mfg/model and uses a transition time of 0x0001 for
the Sengled Element Plus (value in code is in little endian byte format)
and continues to use a value of 0x0000 for everything else.

https://smartthings.atlassian.net/browse/DVCSMP-3163