Adding changeable icons back to zwave-relay and momentary-button-tile DTHs
In a recent update to change these DTHs to use multi-attribute tiles the option to change the device icon for them was lost. This has caused a few users to not be able to change icons on devices that they could before (PROB-1904). This change adds back the ability to change the icons for these two device types.