Rewrite of its-too-cold.groovy
The prior version of its-too-cold does not bother to shut off the appliance after the desired temperature is set. This commit represents a rewrite that handles both starting up, shutting down, and also allowing users to pick interval on which they want to be reminded of a given cold situation.