ICP-2367 Yale Locks send lock operations before notification events (#2627)
* ICP-2367 Yale Locks send lock operations before notification events

We need to enforce the order of notification and lock in order to get the most information possible for the lock/unlock events. This attempts to delay the lock command event until after the notification event. Because if isStateChange, the lock event will actually essentially get ignored as long as the notification event comes.

This should delay the calling of createEvent on the doorLockOperation data so that the isStateChange flag should be set how we expect it to be.

Cleaning up some of the code and adding a debug log for this rare event.