fix status, result, and state not updating
The logic used to see/parse the switch on/off status will not work with modern versions of Sonoff-Tasmota.
This fix has been tested and works against Sonoff-Tasmota v5.10, v5.11, and v5.11.1

Perhaps the format changed in the past and the groovy was not updated.
Do you know what caused you to use code like `def STATUS_PREFIX = "STATUS = "`? That never occurs in the body in the modern tasmota.