Fixed a problem that forced the user settings interface to reset certain selections each time the page would be refreshed by better separating sections, pages and menus.
Fixed a problem with the virtual thermostats not turning on under certain unique circumstances due to an overlapping of motion, modes and contacts events. 

Fixed an issue with the extra-device separately controlled by a separate contact sensor (bed sensor) not turning off when location was not in the proper mode.