WWST-202, DVCSMP-3666 Refactor Ecobee service manger WWST-202 - Refactor Ecobee service manger to not use DTH polling (being depreciated) Move device polling to be controlled by the Service Manager (Connect) app and implement device health in Switch, Thermostat and Sensor. Addition of event subscriptions can't be done until Ecobee releases an API that allows for that.
DVCSMP-3666 - Ecobee (Connect) not able to re-authenticate ST subscription
If, for some reason the authorization grant, token or credentials are invalid,
expired or revoked, Ecobee (connect) may still try to poll the cloud and keep trying to refresh authToken.
For a user to resolve this the ST Ecobee integration needs to re-authenticate the ST subscription in the
Ecobee cloud, however as the connect app skips the login page if authToken exits, there's no check to
see if the token is valid, there's no way for the user to re-authenticate.

Also, changed use of atomicState to state to reduce possible data corruption and added a mechanism
to stop and restart polling when user launches the SM to add/remove devices to avoid data corruption