Ecobee throws groovy.lang.StringWriterIOException: java.io.IOException
When Ecobee (connect) receives HttpResponseException with status 400 a "java.io.IOException: Stream closed"
is thrown while accessing the error object. As a workaround a new string object is created using toString() on
the response data.

Also changing refresh of authToken to only request refresh of the current used and to handle the new token
when the method returns rather than have authToken call the method again, which can cause recursion.