run evaluate() when user changes the settings.
The control loop needs to be updated whenever the temperature sensor reading changes or when the user changes the setpoint.  The temperatureHandler() event handler takes care of updating the control loop when the sensor reading changes, but we need to also update the control loop when the user changes the setpoint.  Running evaluate() in the installed() function body takes care of updating the control loop when the user changes the setpoint.