DVCSMP-3572 Z-Wave Multilevel devices cannot express 100% (#2944)
* DVCSMP-3572 Z-Wave Multilevel devices cannot express 100%

This just parses a value of 99 as 100. Because of how the Z-Wave spec is written 99=100%, 1<1%, but we were treating it as 1-to-1. This is the least painful fix. Exact mapping would with rounding would mean setting any value over 50 would map to that level + 1. Here that only happens at 99.

* DVCSMP-3572 Z-wave multilevel devices cannot express 100%

there's a bug in iOS where the levelSliderControl tile has its default range set to 0-99