CHAD-2137 Remove duplicate events being generated on some z-wave dimmers (#3103)
There has been a complaint about some Z-Wave dimmers producing duplicate
events. The root cause of this issue was creating an event on both the
command and the report. These duplicate events were not being caught by
the duplicate filter because one of the events contained a unit of `%`
while the other was unitless.