After reconsidering we will be removing the event from the entirely (#3104)
We were going to rely on the duplication filter but have shown that
is not going to be a viable solution. The best way to handle this
will just be to remove the event creation from the cloud DTH's
command side. The event will be generated by the hub when a
report event is recieved from the device.