Remove the Level Control cluster from the Samjin Outlet fingerprint
The new firmware is removing the Level Control cluster from the Simple
Descriptor since it doesn't actually allow for control over the  level of
the outlet load. Even so, the correct DTH will still be chosen for devices
that have the old firmware with the Level Control cluster or the new firmware
that doesn't have it.

https://smartthings.atlassian.net/browse/SMJN-164