Create zwave-rgbw-controller.groovy
Creation of a new "zwave-rgbw-controller" based on the "zwave-rgbw-bulb" code (#3628).  The only difference is on a RGBW controller there is no color temperature control and the white channel is on a standard dimmer. The modifications for the white channel/dimmer were based on the existing Fibaro RGBW controller. @greens 

(This replaced a closed pull request that was outdated with recent changes in https://github.com/SmartThingsCommunity/SmartThingsPublic/pull/3633)
(This replaced a