Zigbee Multi Switch: Fix child device naming
The Chinese translation for the join name is "智能墙面开关 三星智家(T21W2Z) 1"
which wasn't working because of the "1" in the join name. I changed it to
just replace the last character with the switch number. This obviously
won't work if the switch number is not at the end or if there are more
than 9 switches.

https://smartthings.atlassian.net/browse/WWST-1485