null: The SmartApp is used to help ensure a door stay locked with no user intervention.
This SmartApp monitors the specified "Open/Closed Sensor" and "Lock": for events.  

If an event shows the door is "closed" or the lock is "unlocked", after N seconds 
the SmartApp will lock the Lock. 

If, during that N seconds, you open the door or lock it, the auto-lock will not be performed.