Hubitat report fix, previous color survives reboot, program settings
Previous color is now kept in EEPROM if the user has "Previous" set as the default color and a program is not running. This will prevent excessive writes to EEPROM.
Hubitat wasn't parsing unsolicited reports correctly because the packet did not have "content length" specified. 
When a program finishes, the controller can return to it's color before the program started, off, or the last color in the program.