[CHAD-3119] Unify Command Class Version Mapping when Parsing Z-Wave Messages
When parsing a Z-Wave message received from devices, a given DTH would need to map that message to a specific version of a command class as different versions of the command class may have different fields and require special parsing. The initial design of the Z-Wave messages could possibly include the encapsulation the message was received with. However, starting in 0.25.X Hub Firmware Release, the hub started removing the encapsulation the message was received with in preparation to handle S2 devices. The encapsulation removal allows most DTHs to support S2 seamlessly without having to worry about encapsulation.

However, with removal of encapsulation of the received messages, we also need to ensure that all the mapping used for encapsulated messages are also used in `zwave.parse()` method as well, and as such, the changes in this PR is to ensure SmartThings DTHs have the same source of command class version mapping regardless of encapsulation.

For more information, please visit: https://community.smartthings.com/t/z-wave-dths-changes-to-encapsulation-handling-starting-in-0-25-x-hub-firmware-release/159433