Aeotec Wallmote: Fix issue syncing components to OCF
By creating the components in 'installed' we avoid a platform bug with the
syncing of components that are created after the device is created (aka in
'updated'). Before this change was made I wasn't seeing all the components
on the device view in the new SmartThings app and after this change I did.

https://smartthings.atlassian.net/browse/WWST-2353
https://smartthings.atlassian.net/browse/WWST-2350