Publish device handlers
The Rooms Beautiful company has 2 ZigBee products that can work with Smartthings. The Curtain device can do open/close, as well as stop/pause functions to control a curtain motor. The Range Extender is allows for range extension for ZigBee devices - it can report how many devices it currently extends the range for.