Zigbee/ZLL DTHs: Added generic profile/deviceId/cluster fingerprints
This change adds/updates the generic fingerprints to a number of the Zigbee
and ZLL DTHs. These generic fingerprints are based on the profile and device
id and the set of clusters used by the DTH. This should increase the number
of devices that join using the most appropriate DTH even though they don't
have a manufacturer/model specific fingerprint. The goal is to provide a
better user experience for users who devices that have not yet gone through
WWST certification but still function OK with one of our existing DTHs.

https://smartthings.atlassian.net/browse/CHAD-3725