Child devices of hub connected devices not being added to room on join
Some hub connected DTHs were not setting the hubId when creating child
devices. This was resulting in the child devices not being shown in the
join screen in the app which had the side effect that they were not being
added to the room the user selected when enabling join.

https://smartthings.atlassian.net/browse/ICP-11101