Removed the unused "enrollResponse" custom command from all DTHs
With the exception of the Tyco Door Window Sensor DTH, all these DTHs
declared support for the custom command "enrollResponse" but none of them
implemented that command. That's likely because in the past they did
implement this command, but then we updated the DTH to use the Zigbee library
version instead but forgot to remove the command.

https://smartthings.atlassian.net/browse/CHAD-3850