[ONEAPP-18302] Z-Wave Siren: Fix infinite loop in initialization
If a Z-Wave siren using this DTH fails to initialize correctly or drops offline, the DTH attempts will generate 3 or more commands every 12 seconds attempting to initialize the device correctly. This leads to the hub's buffer overflowing since it will exhaust its resources attempting to communicate the messages that are bound to fail leading to the eventual queue becoming full and dropping messages.

As a result, I have chosen to cap the attempt to initialize the device after 3 attempts.

https://smartthings.atlassian.net/browse/ONEAPP-18302