CHAD-4375 Check if child devices exist before creating in EcoSmart Remote DTH
The original implementation of the EcoSmart remote DTH did not check if
child devices existed before attempting to create them. This would result
in duplicate child devices being created which would end up resulting in an
error/crash. This manifested as being unable to change the DTH from
self-published to official releases of this specific DTH.