[WWST-6137, WWST-6155, WWST-6142, WWST-6169, WWST-6186, WWST-6192, WWST-6197] EZEX switch fingerprints patch.
EZEX device raw data not include the manufacturer information.
So. This patch have remove the manufacturer field of EZEX finger print.

by WWST-6197 below comment.
======================================================================
Tom Manley

The device doesn’t report a manufacturer but the fingerprint specifies the manufacturer as EZEX. This is why it is joining as a Thing. You need to remove the manufacturer from the fingerprint and then it should join correctly.

For reference, here is the join message from Sumo logs showing the blank manufacturer:
...
==========================================