Ozom Smart Siren: Fix functionality when used with SmartThings Home Monitor
This DTH declares support for the "Alarm" capability but it did not include
implementations for either the `both` or `strobe` commands that are part of
that capability. STHM uses the `both` command and since it was missing from
this DTH any device using this DTH could not be used with STHM. The fix is
to implement the missing commands.

https://smartthings.atlassian.net/browse/CHAD-4456