ICP-11631 Refresh Yale lock battery value after battery replacement
Queues up a battery get after we see the battery changed notification from the lock. Previous implementation relied on sending gets whenever the lock received a new message in parse(), but locks run locally so this happend very rarely. New approach is to query on a regular timer until we get an update.