Schwaiger ZHS01 Dimmer
Schwaiger ZHS01 Dimmer is not supported. It is possible to include as Z-Wave Dimmer Switch Generic. But if you use the wall switch than the status would not displayed in the App. Also if try other dimmer devices on the web interface.
Maybe ist would be possible to fix the problem.
Thank You
Marcel