Zigbee RGBW Bulb: Fix the color temperature range for innr bulbs
These bulbs were using the default UI Metadata for the DTH which supports a
color temperature range of 2700 to 6500 but these bulbs actually support a
range of 1800 to 6500.

Note: this fix will only apply to newly joined devices that match one of the
modified fingerprints. Existing devices will not automatically update to the
new UI Metadata due to the way fingerprint-specific UI metadata is handled.

https://smartthings.atlassian.net/browse/CHAD-4704