Changed arrival sensor beep command endpoint
The arrival sensor beep commands were being send with a source endpoint
of 2. The V3 doesn't support that endpoint and as a result the beep
commands were failing to be sent to the sensor. This fix changes the
source endpoint to be 1 for the beep commands. That endpoint is
supported by both the V2 and V3.

https://smartthings.atlassian.net/browse/CHAD-4825