Local sync
Syncing this with my local repository. Updated to request a battery report if it hasn't received on in over 24 hours, but this device does not wake up unless you press the W button on the back. The device also sends battery reports when it boots up (so when you change out the batteries).