ICP-10928 - Zigbee RGBW Bulb: Fix color temperature value on initial join, code refactoring. (#30421)
Set the color temperature to 5000K on initial join. 5000K was chosen because it is one of the values that is doesn't change when converted to mireds and back again and because it is in the supported range of all the devices that have fingerprints in this DTH.

The code was also refactored to make more use of the Zigbee library.