CHAD-5290 Move Schlage BR469ZP fingerprint (#45988)
Because this device reports door lock operation events prior to notification events, lock code events are generated without the necessary user code data. This was causing user-code based automations to not fire. Since this device executes locally, it would have required a high-risk fix in hubcore. Rather than removing the device entirely, we are moving it to a DTH where user codes would not be accessible for the purposes of creating automations.