New device handler for the Lidl/Silvercrest Motion Detector (Rebranded Tuya Motion detector)
Lidl stores now sell a rebranded range of Tuya Smart Life products. They use their own brand, Silvercrest. The light bulbs etc seem to work fine with ST, but the Motion Detection Sensor did not have a Device handler. 
This Device handler is a modified version of the Smartthings Motion detection DH and gives the basic functionality of identifying the device and adding it to the network, detecting motion and show battery state.