BUG-3164 Overlapping lock events
Lock events for devices for which we delay the sending of lock operation reports that happen too close together will cause incorrect state to temporarily be displayed.

Overwriting the scheduled function call responsible for these events should be enough to essentially drop the "bad" event on the floor so it never has a chance to overwrite good state.