Update aeotec-wallmote.groovy
We test the device "Minoston Remote Control" and find that the deviceJoinName is not take effect. It is always showed as "Aeotec Wallmote", the DTH name. Another brand is showed, wich is not acceptable. So I have to create another DTH and start a new PR(https://github.com/SmartThingsCommunity/SmartThingsPublic/pull/78439).
I will cancel the PRs if your guys can fix it soon. 
Thanks!